% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullcalendar.R
\name{fullcalendar}
\alias{fullcalendar}
\title{Display a FullCalendar widget}
\usage{
fullcalendar(data = NULL, settings = list(), width = NULL,
  height = NULL, elementId = NULL)
}
\arguments{
\item{data}{a data frame with one row per event where the columns needs
to follow the naming conventions by FullCalendar shown here: https://fullcalendar.io/docs/event_data/Event_Object/ .
This data frame will then be translated into a json array by means of \code{jsonlite::toJSON}. The required
columns are \code{title}, a string, and \code{start} as string parseable as a date, for example, "2017-02-28".}

\item{settings}{A list of settings for the FullCalendar.
See here for the available settings: https://fullcalendar.io/docs/display/ .}

\item{width}{Fixed width for widget (in css units). The default is NULL, which results in intelligent automatic sizing based on the widget's container.}

\item{height}{Fixed height for widget (in css units). The default is NULL, which results in intelligent automatic sizing based on the widget's container.}

\item{elementId}{Use an explicit element ID for the widget (rather than an automatically generated one). Useful if you have other JavaScript that needs to explicitly discover and interact with a specific widget instance.}
}
\description{
This function creats a FullCalendar htmlwidget that will show up
in the viewer pane, if you're running Rstudio, or in your Rmarkdown
document.
}
\examples{
data = data.frame(title = paste("Event", 1:3),
                 start = c("2017-03-01", "2017-03-01", "2017-03-15"),
                 end = c("2017-03-02", "2017-03-04", "2017-03-18"),
                 color = c("red", "blue", "green"))
fullcalendar(data)

}
